pyproject.toml
[project]
name = "tds-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Ayush Jain",email = "ayush1997jain@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "typer[all] (>=0.16.0,<0.17.0)",
    "rich (>=14.0.0,<15.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlsxwriter (>=3.2.5,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "openai (>=1.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
tds-app = "tds_app.cli:main"




[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

